refactor(minesweeper) : package 구성을 변경
