--- src/main/kotlin/domain/Mine.kt
@@ -1,5 +1,0 @@
-package domain
-
-data class Mine(
-    val position: Position
-)
