+++ src/main/kotlin/domain/mine/Mine.kt
@@ -1,0 +1,7 @@
+package domain.mine
+
+import domain.position.Position
+
+data class Mine(
+    val position: Position
+)
