+++ src/main/kotlin/domain/mine/Mines.kt
@@ -1,4 +1,6 @@
-package domain
+package domain.mine
+
+import domain.position.Position
 
 class Mines(val values: List<Mine>) {
     fun isMine(position: Position): Boolean {
