+++ src/main/kotlin/domain/position/PositionIdFactory.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.position
 
 interface PositionIdFactory {
     fun positionIds(count: Int, positionIdMax: Int): List<PositionId>
