+++ src/main/kotlin/domain/position/RandomPositionIdFactory.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.position
 
 import kotlin.random.Random
