+++ src/test/kotlin/domain/MineFactoryTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.mine.MineFactory
 import model.GameData
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
