+++ src/test/kotlin/domain/PositionIdTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.position.PositionId
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
